Changed Landcover for calculation
Land cover changed from full area to just the landcover within the open street map (OSM) residential area. This was done to make the EPA results more comprable to Tobler's functions. Note that this will increase the apparent density. This is because the tool will "stuff" the population outside the residential zones into it. We are currently checking this over.